# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "High-Performance Anime Streaming API built with Go\nA blazing-fast, production-ready RESTful API for anime streaming platforms\n\n**Features:**\n- Complete anime data (episodes, metadata, streaming sources)\n- Advanced search and filtering\n- Schedule management and notifications\n- Character and voice actor information\n- User watchlist support\n- Optimized caching with Redis\n\n**Disclaimer:**\nThis API does not store any files. It only links to media hosted on 3rd party services.\nThis API is explicitly made for educational purposes only.",
        "title": "Jutsu API",
        "contact": {
            "name": "Jutsu API Support",
            "url": "https://github.com/mnuddindev/jutsu-api",
            "email": "support@jutsu-api.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/character/list/{id}": {
            "get": {
                "description": "Returns a paginated list of characters with voice actors for a specific anime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Characters"
                ],
                "summary": "Get anime characters",
                "parameters": [
                    {
                        "type": "string",
                        "example": "frieren-beyond-journeys-end-18542",
                        "description": "Anime ID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Characters and voice actors",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to fetch character list",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/character/{id}": {
            "get": {
                "description": "Returns detailed information about a specific character including voice actors and animeography",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Characters"
                ],
                "summary": "Get character details",
                "parameters": [
                    {
                        "type": "string",
                        "example": "asta-340",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character details",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Missing path parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Character not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to fetch character",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/episodes/{id}": {
            "get": {
                "description": "Returns the list of episodes for a specific anime",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anime"
                ],
                "summary": "Get anime episode list",
                "parameters": [
                    {
                        "type": "string",
                        "example": "frieren-beyond-journeys-end-18542",
                        "description": "Anime ID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Episode list",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Missing or invalid path parameter",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to fetch episodes",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/producer/{id}": {
            "get": {
                "description": "Returns paginated anime rows for a producer identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get anime of specific producers",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1",
                        "description": "Producer ID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producer anime list",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Page out of range",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load producer data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/studio/{id}": {
            "get": {
                "description": "Returns paginated anime rows for a studio identifier",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Categories"
                ],
                "summary": "Get anime of specific studios",
                "parameters": [
                    {
                        "type": "string",
                        "example": "1",
                        "description": "Studio ID or slug",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "example": 1,
                        "description": "Page number (must be \u003e= 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Studio anime list",
                        "schema": {
                            "$ref": "#/definitions/handler.CategoryResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Page out of range",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to load studio data",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.CategoryResponse": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "object",
                    "properties": {
                        "data": {},
                        "totalPages": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "handler.ErrorResponse": {
            "description": "Standard error response",
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Error message here"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handler.SuccessResponse": {
            "description": "Standard success response",
            "type": "object",
            "properties": {
                "cached": {
                    "type": "boolean",
                    "example": false
                },
                "results": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
...
}